Access post based on time.
We're doing this using the `since` and `until` parameters of the API. `until` is the current time and `since` is one hour prior to `until`
ALso, we're keeping track of the last post we accessed, i.e, the first element of the array containing the posts, by recording its id. So, the moment we encounter this id, we can skip all posts from that point forward and that post itself.