Thoroughly document code
Documented the entire code so that the documentation can be generated by
running the command ```$ jsdoc src/services/*.js src/index.js```

Be sure to install jsdoc by running ```$ npm i -g jsdoc```

The documentation is then saved inthe ```out/``` folder in the root directory
of the project which is ignored by Git.
In the ```out/``` folder, open ```index.html``` to view the documentation.