Migrate to Firebase Realtime DB
Completed the migration of saving all the posts to Firebase Realtime DB.
The workings are the same at high level.

Internally, we're running the save posts query for every single post in a batch.
This is because, each save query saves a record with a unique key. That means,
when a bunch posts in a list are saved using one save query, the posts get saved
as nested records under one uniue key, which serves as the ID for each record.

To test for now, just call the functions in src/services/dbConnect.js

Be sure to take the service account credentials from me. The Firebase console
generates keys only once and it will be a problem if the keys are changed constantly.
The .env is also updated to have the URL to the DB.

Install the new dependencies.